Total balance as default buy amount
